More basic functionality implemented
We can now view a single project page, view a single customer/client page along with their associated projects. Furthermore, we can edit customers and projects as well as delete customers and projects. The next commit will address the issue with the sidebar and it's unreponsiveness as well as the general front end routing.